@@ -1,16 +1,18 @@
 [metadata]
 name = lcov_cobertura
-version = attr: lcov_cobertura.lcov_cobertura.__version__
+version = attr: setuptools_scm.get_version
+description = LCOV to Cobertura XML converter
+url = https://eriwen.github.io/lcov-to-cobertura-xml/
 author = Eric Wendelin
 author_email = me@eriwen.com
 maintainer = Steve Arnold
 maintainer_email = nerdboy@gentoo.org
-description = LCOV to Cobertura XML converter
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://eriwen.github.io/lcov-to-cobertura-xml/
 download_url = https://raw.githubusercontent.com/eriwen/lcov-to-cobertura-xml/master/lcov_cobertura/lcov_cobertura.py
 license = Apache License, Version 2.0
+license_expression = Apache-2.0
+license_files = LICENSE
 classifiers =
     Programming Language :: Python
     Intended Audience :: Developers
@@ -21,8 +23,6 @@ classifiers =
     Topic :: Software Development :: Quality Assurance
     Development Status :: 5 - Production/Stable
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
 
@@ -31,10 +31,16 @@ keywords =
     cobertura
 
 [options]
-python_requires = >=3.6
-install_requires =
+python_requires = >=3.8
 
-py_modules = lcov_cobertura/lcov_cobertura
+setup_requires =
+    setuptools_scm[toml]
+
+install_requires =
+    importlib-metadata; python_version < '3.8'
+
+packages =
+    lcov_cobertura
 
 [options.entry_points]
 console_scripts =
