@@ -13,15 +13,14 @@
 import os
 import sys
 
-import pkg_resources
-
+if sys.version_info < (3, 8):
+    from importlib_metadata import version
+else:
+    from importlib.metadata import version
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
-# needs an extra path here for namespace module-not-a-package
-# otherwise apidoc generates empty stub pages
-sys.path.insert(0, os.path.abspath('../../lcov_cobertura'))
 
-__version__ = pkg_resources.get_distribution('lcov_cobertura').version
+__version__ = version('lcov_cobertura')
 
 # -- Project information -----------------------------------------------------
 
@@ -86,7 +85,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+#html_static_path = ['_static']
 
 # If true, "(C) Copyright ..." is shown in the HTML footer. Default is True.
 html_show_copyright = False
